remove duplicated code of create and interval API
