+++ rxbp/flowables/createflowable.py
@@ -2,7 +2,7 @@ from rx.core import typing
 
 from rxbp.init.initsubscription import init_subscription
 from rxbp.mixins.flowablemixin import FlowableMixin
-from rxbp.observables.createobservable import CreateObservable
+from rxbp.observables.subscriptionobservable import SubscriptionObservable
 from rxbp.subscriber import Subscriber
 from rxbp.subscription import Subscription
 
@@ -16,7 +16,7 @@ class CreateFlowable(FlowableMixin):
 
     def unsafe_subscribe(self, subscriber: Subscriber) -> Subscription:
         return init_subscription(
-            observable=CreateObservable(
+            observable=SubscriptionObservable(
                 source=self._observer,
                 scheduler=subscriber.scheduler,
                 subscribe_scheduler=subscriber.subscribe_scheduler,
