+++ rxbp/flowables/intervalflowable.py
@@ -3,7 +3,7 @@ from rx.core import typing
 from rxbp.init.initsubscription import init_subscription
 from rxbp.internal.timer import _interval
 from rxbp.mixins.flowablemixin import FlowableMixin
-from rxbp.observables.intervalobservable import IntervalObservable
+from rxbp.observables.subscriptionobservable import SubscriptionObservable
 from rxbp.subscriber import Subscriber
 from rxbp.subscription import Subscription
 
@@ -17,7 +17,7 @@ class IntervalFlowable(FlowableMixin):
 
     def unsafe_subscribe(self, subscriber: Subscriber) -> Subscription:
         return init_subscription(
-            observable=IntervalObservable(
+            observable=SubscriptionObservable(
                 source=_interval(self.period),
                 scheduler=subscriber.scheduler,
                 subscribe_scheduler=subscriber.subscribe_scheduler,
